+++ CHANGELOG.rst
@@ -3,6 +3,12 @@ Changelog
 
 Main changes to Sound Speed Manager:
 
+Version 2025.X.Y (2025-MM-DD)
+-----------------------------
+
+- Added file logging in SSM_DEBUG mode.
+
+
 Version 2025.4.6 (2025-08-17)
 -----------------------------